PoC/slack_deploy_bot/requirements.txt
GitPython == 3.1.43
python-dotenv == 1.0.1 
slack-bolt == 1.18.1 